Finish
Endpoint to get a list of upcoming movies.
The same endpoint should return the list of the next 20 movies as page param is given.

Endpoint to get a list of top rated movies.
The same endpoint should return the list of the next 20 movies as page param is given.

Endpoint to get a specific single movie.
The same endpoint should return all movie related videos with a single request.

Endpoint to get a list of genres.
The same endpoint should return a single genre by id.